requirements.txt
opencv-python
Pillow
hachoir
ffmpeg-python
libtorrent
requests
Pyrogram
python-magic
rarfile
